pyproject.toml
[project]
name = "odata"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.40.55",
    "cachetools>=6.2.1",
    "defusedxml>=0.7.1",
    "fastapi>=0.119.0",
    "google-cloud-bigquery>=3.38.0",
    "google-cloud-secret-manager>=2.25.0",
    "google-cloud-storage>=3.4.1",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
]
